added "sign_as" to deployment command
